Auth progress

+++ frontend/src/components/NotDetay.js
@@ -2,15 +2,23 @@ import React from 'react'
 import { useNotcontext } from '../hooks/useNotContext'
 import moment from 'moment'
 import 'moment/locale/tr'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 export default function NotDetay({not}) { //homede dolduruyoruz
   
     const {dispatch}=useNotcontext()
+    const {kullanici}=useAuthContext();
 
     const handleClick=async ()=>{
         console.log(not._id)
+        if(!kullanici){
+            return
+        }
         const response = await fetch('/api/notlar/'+not._id,{
-            method:'DELETE'
+            method:'DELETE',
+            headers:{
+                'Authorization':`Bearer ${kullanici.token}`
+            }
         })
 
         const json =await response.json()
